feat: create ThemeProvider
